requirements.txt
langchain==0.1.20
langchain-anthropic==0.1.15
langchain-core==0.1.50
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.11.0
redis==5.0.1
